refactor: Improve code structure and readability
