+++ public/site.webmanifest
@@ -1,7 +1,7 @@
 {
-  "name": "Next Template",
+  "name": "Guestbook",
   "short_name": "",
-  "description": "Minimal Next.js template.",
+  "description": "Full stack web application built with everthing new.",
   "theme_color": "#ffffff",
   "background_color": "#ffffff",
   "display": "standalone"
