+++ src/app/sitemap.ts
@@ -1,4 +1,4 @@
-import { siteConfig } from "~/config/site"
+import { siteConfig } from "~/config"
 
 export default async function sitemap() {
   return [
